Separates unit & integration tests
