+++ .github/workflows/unit-tests.yml
@@ -5,10 +5,6 @@ on:
         branches: [main]
     pull_request:
 
-env:
-    INTEGRATION_ENVIRONMENT_ID: ${{ secrets.INTEGRATION_ENVIRONMENT_ID }}
-    INTEGRATION_MANAGEMENT_API_KEY: ${{ secrets.INTEGRATION_MANAGEMENT_API_KEY }}
-
 jobs:
     build:
         runs-on: ubuntu-latest
@@ -19,4 +15,4 @@ jobs:
               with:
                   node-version-file: '.nvmrc'
             - run: npm ci
-            - run: npm run test
+            - run: npm run test:unit
