+++ package.json
@@ -32,7 +32,9 @@
 		"prepublishOnly": "npm run build && npm run push:tags",
 		"publish:beta": "npm publish --tag=next",
 		"build": "npm run biome:check && npm run eslint && npm run clean &&  tsc --p tsconfig.build.json && npm run update:version",
-		"test": "vitest run --config=vitest.config.ts",
+		"test": "npm run test:unit && npm run test:integration",
+		"test:unit": "vitest run --config=vitest-unit.config.ts",
+		"test:integration": "vitest run --config=vitest-integration.config.ts",
 		"eslint": "eslint",
 		"biome:check": "biome check",
 		"biome:fix": "biome check --write",
