+++ tests/unit/retry-policy/retry-policy-max-attempts.spec.ts
@@ -1,10 +1,9 @@
 import { afterAll, describe, expect, it, vi } from "vitest";
-import { getFetchJsonMock } from "../../lib/devkit/test.utils.js";
-import type { HttpResponse } from "../../lib/http/http.models.js";
-import { getDefaultHttpService } from "../../lib/http/http.service.js";
-import type { RetryStrategyOptions } from "../../lib/models/core.models.js";
-import { toRequiredRetryStrategyOptions } from "../../lib/utils/retry.utils.js";
-import { getIntegrationTestConfig } from "../integration-tests.config.js";
+import { getFetchJsonMock } from "../../../lib/devkit/test.utils.js";
+import type { HttpResponse } from "../../../lib/http/http.models.js";
+import { getDefaultHttpService } from "../../../lib/http/http.service.js";
+import type { RetryStrategyOptions } from "../../../lib/models/core.models.js";
+import { toRequiredRetryStrategyOptions } from "../../../lib/utils/retry.utils.js";
 
 const testCases: readonly Required<RetryStrategyOptions>[] = [
 	toRequiredRetryStrategyOptions({
@@ -65,7 +64,7 @@ async function resolveResponseAsync(retryStrategy: Required<RetryStrategyOptions
 		retryStrategy,
 	}).requestAsync({
 		// we need valid url
-		url: getIntegrationTestConfig().urls.baseMapiUrl,
+		url: "https://domain.com",
 		method: "GET",
 		body: null,
 	});
